Split dingo/core/factors.py into the dingo.core.inference package
`dingo.core.factors` held four concerns under a name that fit only one of
them. Split it into a package mirroring `dingo/gw/inference/`:

- `dingo/core/inference/context.py`: the `SamplerContext` protocol
  (implemented by `dingo.gw.inference.context.GWSamplerContext`).
- `dingo/core/inference/steps.py`: `Step`, `Factor`, `Standardization`,
  `FlowFactor`, `DeltaFactor`, `SampleTableFactor`, `Reparametrization`,
  `ProxyOffsetReparam`, `TargetCorrection`, and the step-level helpers.
- `dingo/core/inference/composer.py`: `Stage`, `ChainComposer`, `GibbsBlock`,
  `ComposedSampler`, `chunk_and_concat`.

Definitions are moved verbatim; only module docstrings, imports, and the docs'
API references change.
